requirements.txt
Flask==3.1.0
flask-cors==5.0.0
langchain-openai==0.2.12  
langchain-core==0.3.25  
langgraph==0.2.59 
langgraph-checkpoint-sqlite==2.0.3
faiss-cpu==1.9.0.post1
langchain-huggingface==0.1.2
langchain-community==0.3.16